pyproject.toml
[tool.poetry]
name = "FilterBot"
version = "0.1.0"
description = ""
authors = ["Rasul Taimaskhanov <81189712+taimaskhanov11@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
aiogram = "^2.19"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.0"}
PyYAML = "^6.0"
pydantic = "^1.9.0"
APScheduler = "^3.9.1"
aioredis = {extras = ["hiredis"], version = "^2.0.1"}
action = "^1.4.4"
Telethon = {extras = ["cryptg"], version = "^1.24.0"}

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
